requirements.txt
langchain 
langchain-community 
openai 
chromadb 
pymupdf 
beautifulsoup4 
requests 
fastapi 
uvicorn 
sqlalchemy 
wikipedia-api
newspaper3k
trafilatura
python-dotenv
